Restructure blast directory
