+++ graphsearch/hit.cpp
