+++ graphsearch/hit.h
