+++ graphsearch/queries.h
