+++ graphsearch/query.cpp
